Main: Clear user_input when continuing
read_line() appends to the buffer, so if we have a value that is outside
the 1..TOPIC_COUNT range, it still gets appended to the buffer and then
we continue. What happens is then if we give a valid number after that
invalid number, the buffer contains both numbers. So we need to clear.